Добавлена логика для передачи изображений

+++ src/main/java/ru/skypro/homework/repository/ImageRepository.java
@@ -1,7 +1,10 @@
 package ru.skypro.homework.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import ru.skypro.homework.entity.Comment;
 import ru.skypro.homework.entity.Image;
 
-public interface ImageRepository extends JpaRepository<Image, Long> {
+import java.util.List;
+
+public interface ImageRepository extends JpaRepository<Image, Integer> {
 }
